wait

+++ gateway/src/router.rs
@@ -209,28 +209,36 @@ impl Router {
         to_delete
     }
     fn cleanup_game_states(&mut self) {
-        let to_delete = self.find_dead_game_states();
-        let mut count = 0;
-        for game_id in to_delete {
-            self.game_states.remove_entry(&game_id);
+        if let Some(dur) = Instant::now().checked_duration_since(self.last_cleanup) {
+            if dur.as_millis() > GAME_STATE_CLEANUP_PERIOD_MS.into() {
+                let to_delete = self.find_dead_game_states();
+                let mut count = 0;
+                for game_id in to_delete {
+                    self.game_states.remove_entry(&game_id);
 
-            // ...and make sure no one else can wander
-            // into this game, since it's no longer playable
-            match self.available_games.get_mut(0) {
-                Some(gid) if gid == &game_id => {
-                    self.available_games.pop();
-                    println!("POP");
+                    // ...and make sure no one else can wander
+                    // into this game, since it's no longer playable
+                    match self.available_games.get_mut(0) {
+                        Some(gid) if gid == &game_id => {
+                            self.available_games.pop();
+                            println!("POP");
+                        }
+                        _ => (),
+                    }
+                    count += 1;
                 }
-                _ => (),
-            }
-            count += 1;
-        }
 
-        if count > 0 {
-            println!(
-                "🗑 {} {} {:<8} {} entries",
-                EMPTY_SHORT_UUID, EMPTY_SHORT_UUID, "CLEANUP", count
-            )
+                if count > 0 {
+                    println!(
+                        "🗑 {} {} {:<8} {} entries",
+                        EMPTY_SHORT_UUID, EMPTY_SHORT_UUID, "CLEANUP", count
+                    )
+                }
+            } else {
+                println!("wait")
+            }
+        } else {
+            println!("wait")
         }
     }
